shinyapp/requirements.txt
shiny
python-dotenv
tokenizers
anthropic

# For emails
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
markdown
pandas
requests
